handle JWT Authorization

+++ PPMToolFullStack/src/main/java/io/billal/ppmtool/services/ProjectService.java
@@ -3,11 +3,16 @@ package io.billal.ppmtool.services;
 
 import io.billal.ppmtool.domain.Backlog;
 import io.billal.ppmtool.domain.Project;
+import io.billal.ppmtool.domain.User;
 import io.billal.ppmtool.exceptions.ProjectIdException;
+import io.billal.ppmtool.exceptions.ProjectNotFoundException;
 import io.billal.ppmtool.repositories.BacklogRepository;
 import io.billal.ppmtool.repositories.ProjectRepository;
+import io.billal.ppmtool.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.security.Principal;
 
 @Service
 public class ProjectService {
@@ -18,9 +23,24 @@ public class ProjectService {
     @Autowired
     private BacklogRepository backlogRepository;
 
-    public Project saveOrUpdateProject(Project project){
+    @Autowired
+    private UserRepository userRepository;
+
+    public Project saveOrUpdateProject(Project project, String username){
+
+        if(project.getId() != null){
+            Project existProject = projectRepository.findByProjectIdentifier(project.getProjectIdentifier());
+            if(existProject != null &&(!existProject.getProjectLeader().equals(username))){
+                throw new ProjectNotFoundException("Project not found in your account");
+            }else if(existProject == null){
+                throw new ProjectNotFoundException("Project with ID: '" + project.getProjectIdentifier() + "' cannot be updated because is doesn't exist!");
+            }
+        }
 
         try{
+            User user = userRepository.findByUsername(username);
+            project.setUser(user);
+            project.setProjectLeader(user.getUsername());
             project.setProjectIdentifier(project.getProjectIdentifier().toUpperCase());
 
             if(project.getId() == null){
@@ -40,23 +60,32 @@ public class ProjectService {
 
     }
 
-    public Project findProjectByIdentifier(String projectId){
+    public Project findProjectByIdentifier(String projectId, Principal principal){
 
         Project project = projectRepository.findByProjectIdentifier(projectId.toUpperCase());
+
         if(project == null){
             throw new ProjectIdException("ProjectID '" + projectId.toUpperCase() + "' doesn't exist");
         }
+
+        if(!project.getProjectLeader().equals(principal.getName())){
+            throw new ProjectNotFoundException("Project '" + projectId.toUpperCase() + "' not found in your account");
+        }
+        
         return project;
     }
 
-    public Iterable<Project> findAllProjects(){
-        return projectRepository.findAll();
+    public Iterable<Project> findAllProjects(String username){
+        return projectRepository.findAllByProjectLeader(username);
     }
 
-    public void deleteProjectByIdentifier(String projectId){
+    public void deleteProjectByIdentifier(String projectId, Principal principal){
         Project project = projectRepository.findByProjectIdentifier(projectId.toUpperCase());
         if (project == null){
             throw new ProjectIdException("Cannot delete this ID '" + projectId.toUpperCase() + "' the ID does not exist");
+        }
+        if(!project.getProjectLeader().equals(principal.getName())){
+            throw new ProjectNotFoundException("Project not found in your account");
         }
         projectRepository.delete(project);
     }
